repro_342: explicit double casts for the commit-time print (clang-18)
The Linux GCP build (clang-18) errored on -Wimplicit-int-float-conversion at
repro_342.c:380 -- the time_t/long members of the timespec diff were
implicitly converted to double (int64 -> double can lose precision), forcing
the GCP pass to build with -DSTM_WERROR=OFF. Add explicit (double) casts on
the tv_sec/tv_nsec diffs (matching tharness.c:62-63, already cast) so the
next sanitizer pass builds -Werror-clean and turnkey. No behavior change;
repro_342 builds + links.

+++ v2/tests/repro_342.c
@@ -377,7 +377,8 @@ static int wedge_load(stm_fs *fs, uint64_t dir)
     clock_gettime(CLOCK_MONOTONIC, &t0);
     stm_status crc_ = stm_fs_commit(fs);
     clock_gettime(CLOCK_MONOTONIC, &t1);
-    double ms = (t1.tv_sec - t0.tv_sec) * 1e3 + (t1.tv_nsec - t0.tv_nsec) / 1e6;
+    double ms = (double)(t1.tv_sec - t0.tv_sec) * 1e3
+              + (double)(t1.tv_nsec - t0.tv_nsec) / 1e6;
     printf("wedge: commit rc=%d in %.0f ms\n", (int)crc_, ms);
 
     uint64_t ino2 = 0;
